Done refactoring + wishlist styling
